Complete Part 4 Exercises 4.15-4.23: Token Authentication, bcrypt, jwt

+++ part4/bloglist/controllers/blogs.js
@@ -1,13 +1,20 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({})
     response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response, next) => {
     const body = request.body
+    const user = request.user
+
+    if (!user) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
 
     if (!body.title || !body.url) {
         return response.status(400).end()
@@ -17,19 +24,37 @@ blogsRouter.post('/', async (request, response, next) => {
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes || 0
+        likes: body.likes || 0,
+        user: user.id
     })
 
     try {
         const savedBlog = await blog.save()
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
+
         response.status(201).json(savedBlog)
     } catch (error) {
         next(error)
     }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response, next) => {
     try {
+        const user = request.user
+        if (!user) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+
+        const blog = await Blog.findById(request.params.id)
+        if (!blog) {
+            return response.status(404).end()
+        }
+
+        if (blog.user.toString() !== user.id.toString()) {
+            return response.status(403).json({ error: 'permission denied' })
+        }
+
         await Blog.findByIdAndDelete(request.params.id)
         response.status(204).end()
     } catch (error) {
